test: custom domain with proposals
