+++ src/tests/specs/console/console.cdn.custom-domains.spec.ts
@@ -8,7 +8,7 @@ import { inject } from 'vitest';
 import { CONTROLLER_ERROR_MSG } from '../../constants/console-tests.constants';
 import { uploadFile } from '../../utils/cdn-tests.utils';
 import {
-	adminCustomDomainsTests,
+	adminCustomDomainsTests, adminCustomDomainsWithProposalTests,
 	anonymousCustomDomainsTests
 } from '../../utils/custom-domains-tests.utils';
 import { CONSOLE_WASM_PATH } from '../../utils/setup-tests.utils';
@@ -51,57 +51,6 @@ describe('Console > Cdn > Custom Domains', () => {
 
 		adminCustomDomainsTests({ actor: () => actor });
 
-		it('should not delete asset ic-domains when deleting all assets', async () => {
-			const { http_request, init_proposal, submit_proposal, commit_proposal } = actor;
-
-			const [proposalId, _] = await init_proposal({
-				AssetsUpgrade: {
-					clear_existing_assets: toNullable(true)
-				}
-			});
-
-			await uploadFile({ proposalId, actor });
-
-			const [__, proposal] = await submit_proposal(proposalId);
-
-			await commit_proposal({
-				sha256: fromNullable(proposal.sha256)!,
-				proposal_id: proposalId
-			});
-
-			const { status_code } = await http_request({
-				body: [],
-				certificate_version: toNullable(),
-				headers: [],
-				method: 'GET',
-				url: '/.well-known/ic-domains'
-			});
-
-			expect(status_code).toEqual(200);
-		});
-
-		it('should throw error if try to upload ic-domains', async () => {
-			const { init_asset_upload, init_proposal } = actor;
-
-			const [proposalId, _] = await init_proposal({
-				AssetsUpgrade: {
-					clear_existing_assets: toNullable()
-				}
-			});
-
-			await expect(
-				init_asset_upload(
-					{
-						collection: '#dapp',
-						description: toNullable(),
-						encoding_type: [],
-						full_path: '/.well-known/ic-domains',
-						name: 'ic-domains',
-						token: toNullable()
-					},
-					proposalId
-				)
-			).rejects.toThrow('/.well-known/ic-domains is a reserved asset.');
-		});
+		adminCustomDomainsWithProposalTests({ actor: () => actor });
 	});
 });
